fix: gallery loading by switching to pure Vue
