+++ packages/docs-site/blog/v3.md
@@ -11,7 +11,7 @@ authors:
 <script setup>
 import Author from "../.vitepress/theme/Author.vue";
 import registry from "@penrose/examples/dist/registry.js";
-import Gallery from '../src/components/GalleryWrapper.vue'
+import Gallery from '../src/components/Gallery.vue'
 import {useData} from "vitepress"
 const examples = ["spectral-graphs/examples/hypercube", "spectral-graphs/examples/hexagonal-lattice", "timeline/penrose", "walk-on-spheres/SignedAngleOutside"]
 const trios = [...registry.entries()]
