--- packages/docs-site/src/components/GalleryWrapper.vue
@@ -1,23 +1,0 @@
-<script setup lang="ts">
-import { withBase, useData } from "vitepress";
-import { defineAsyncComponent } from "vue";
-
-const Gallery = defineAsyncComponent(async () => {
-  const { applyPureReactInVue } = await import("veaury");
-  const { Gallery } = await import("@penrose/components");
-  return applyPureReactInVue(Gallery);
-});
-const props = defineProps<{ trios: string[] }>();
-
-const link = withBase(`/try/index.html`);
-</script>
-
-<template>
-  <div style="width: 100%">
-    <Gallery
-      :trios="props.trios"
-      :ideLink="link"
-      :dark="useData().isDark.value"
-    />
-  </div>
-</template>
